// src/store/medication_com/types.ts
export interface IMedicationPayload {
  name: string;
  countryId: number;
  classificationId: string;
  classificationName: string;
  brandId: string;
  code?: ICodeDetailsPayload;
  brandName: string;
  dosageFormId: string;
  dosageFormName: string;
  id?: string;
}

export interface ICodeDetailsPayload {
  code: string;
  url: string;
}
// ...
export interface IDeleteMedicationRequestPayload {
  id: string;
  tenantId: string;
}